requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# NLP Libraries
spacy==3.7.2
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
scikit-learn==1.3.2
textblob==0.17.1
nltk==3.8.1

# Document Processing
PyMuPDF==1.23.5
python-docx==1.1.0
python-multipart==0.0.6
aiofiles==23.2.1

# Vector Database
chromadb==0.4.15

# Redis for caching and message queue
redis[hiredis]==5.0.1
aioredis==2.0.1

# Additional utilities
numpy==1.24.3
pandas==2.0.3
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

